Overhaul of Items system
Changed so that items are stackable. Also you must interact with the item to pick it up. Stats are displayed too.

+++ Glowkeeper/Assets/Scripts/EnemyScripts/EnemyScript.cs
@@ -51,7 +51,7 @@ public class EnemyScript : MonoBehaviour
         }
     }
 
-    public void EnemyDamage(int amount)
+    public void EnemyDamage(float amount)
     {
         enemyHealth -= amount;
